pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "codagent"
version = "0.5.0"
description = "Production utilities for Python LLM agents — node wrappers (retry/timeout/cache), tool decorators (validate/circuit/rate-limit), observability (cost/step/trace), behavior contracts. Works with LangGraph, LangChain, LlamaIndex, CrewAI, AutoGen, DSPy, Pydantic AI, OpenAI, Anthropic."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "rlaope" }]
keywords = [
    "llm", "agents", "guardrails", "behavior-contracts",
    "claude-code", "cursor", "copilot", "openai", "anthropic",
    "langgraph", "langchain", "harness", "karpathy",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Topic :: Software Development :: Libraries",
]
dependencies = []

[project.optional-dependencies]
openai = ["openai>=1.0"]
anthropic = ["anthropic>=0.30"]
langchain = ["langchain-core>=0.2"]
langgraph = ["langgraph>=0.2"]
llamaindex = ["llama-index-core>=0.11"]
pydantic-ai = ["pydantic-ai>=0.0.1"]
crewai = ["crewai>=0.1"]
autogen = ["pyautogen>=0.2"]
dspy = ["dspy-ai>=2.5"]
guardrails-ai = ["guardrails-ai>=0.5"]
nemo = ["nemoguardrails>=0.10"]
test = ["pytest>=8.0"]

[project.scripts]
codagent = "codagent.cli:main"

[project.urls]
Homepage = "https://github.com/rlaope/codagent"
Issues = "https://github.com/rlaope/codagent/issues"
Repository = "https://github.com/rlaope/codagent"

[tool.setuptools.packages.find]
include = ["codagent*"]

[tool.setuptools.package-data]
"codagent.observability" = ["prices.json"]

[tool.pytest.ini_options]
testpaths = ["tests"]
